rename filenames to match action name
