+++ src/server/components/dashboard/listsItems/item/update/actionHandlers/index.ts
@@ -1,0 +1,3 @@
+import { Action, ActionRequestHandler } from "../types";
+
+export const actionHandlers: Record<Action, ActionRequestHandler> = {};
