+++ src/server/components/dashboard/listsItems/item/update/actionHandlers/publish.ts
@@ -3,34 +3,12 @@ import { EventJsonData, User } from "server/models/types";
 import { logger } from "server/services/logger";
 import { prisma } from "server/models/db/prisma-client";
 import { ListItemEvent } from "@prisma/client";
-import { togglerListItemIsPublished, update } from "server/models/listItem";
+import { togglerListItemIsPublished } from "server/models/listItem";
 import { createListSearchBaseLink } from "server/components/lists";
 import { getListItemContactInformation } from "server/models/listItem/providers/helpers";
 import serviceName from "server/utils/service-name";
 import { sendDataPublishedEmail } from "server/services/govuk-notify";
-
-export async function listItemUpdateController(req: Request, res: Response): Promise<void> {
-  const userId = req.user!.id;
-  const listItem = res.locals.listItem;
-  const { listItemUrl, listIndexUrl } = res.locals;
-
-  try {
-    await handleListItemUpdate(listItem.id, userId);
-
-    if (userId === undefined) {
-      req.flash("errorMsg", "Unable to perform action - user could not be identified");
-      return res.redirect(listItemUrl);
-    }
-
-    req.flash("successBannerTitle", `${listItem.jsonData.organisationName} has been updated and published`);
-    req.flash("successBannerHeading", "Updated and published");
-    req.flash("successBannerColour", "green");
-    return res.redirect(listIndexUrl);
-  } catch (error: any) {
-    req.flash("errorMsg", `${listItem.jsonData.organisationName} could not be updated. ${error.message}`);
-    return res.redirect(listItemUrl);
-  }
-}
+import { update } from "server/components/dashboard/listsItems/item/update/actionHandlers/update";
 
 export async function handleListItemUpdate(id: number, userId: User["id"]): Promise<void> {
   logger.info(`${userId} looking for ${id} to update`);
